# [csharp] Optimization message id generator performance.