--- a/csharp/rocketmq-client-csharp/IUtilities.cs
+++ b/csharp/rocketmq-client-csharp/IUtilities.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Org.Apache.Rocketmq
+{
+    internal interface IUtilities
+    {
+        byte[] GetMacAddress();
+        int GetProcessId();
+        string ByteArrayToHexString(ReadOnlySpan<byte> bytes);
+    }
+}